Present companion pairing card from the bottom, drop top banner
